Redesigned system into modularity to improve maintenance and expansion
