+++ DevSync.PocPro/accounts/src/DevSync.PocPro.Accounts.Api/DevSync.PocPro.Accounts.Api.http
@@ -1,0 +1,6 @@
+@DevSync.PocPro.Accounts.Api_HostAddress = http://localhost:5104
+
+GET {{DevSync.PocPro.Accounts.Api_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
