+++ DevSync.PocPro/accounts/src/DevSync.PocPro.Accounts.Api/GlobalUsings.cs
@@ -1,0 +1,9 @@
+// Global using directives
+
+global using System.Collections.ObjectModel;
+global using DevSync.PocPro.Accounts.Api.Features.Tenants.Domain;
+global using DevSync.PocPro.Shared.Domain.Abstractions;
+global using DevSync.PocPro.Shared.Domain.Enums;
+global using DevSync.PocPro.Shared.Domain.Exceptions;
+global using FastEndpoints;
+global using FluentResults;
