+++ DevSync.PocPro/accounts/src/DevSync.PocPro.Accounts.Api/Program.cs
@@ -1,0 +1,7 @@
+using DevSync.PocPro.Accounts.Api.DI;
+
+var builder = WebApplication.CreateBuilder(args);
+
+var app = builder.AddServices().AddPipeline();
+
+app.Run();
