+++ DevSync.PocPro/shared/DevSync.PocPro.Shared.Domain/Enums/PermissionType.cs
@@ -1,0 +1,6 @@
+namespace DevSync.PocPro.Shared.Domain.Enums;
+
+public enum PermissionType
+{
+    MANAGE_PERMISSIONS = 1
+}
