+++ DevSync.PocPro/shared/DevSync.PocPro.Shared.Domain/Exceptions/DomainExceptions.cs
@@ -1,0 +1,3 @@
+namespace DevSync.PocPro.Shared.Domain.Exceptions;
+
+public class DomainExceptions(string message) : Exception(message);
