--- DevSync.PocPro/src/core/DevSync.PocPro.Domain/Abstractions/IEntity.cs
@@ -1,16 +1,0 @@
-namespace DevSync.PocPro.Domain.Abstractions;
-
-public interface IEntity<T> : IEntity
-{
-    public T Id { get; set; }
-}
-
-public interface IEntity
-{
-    public DateTimeOffset? CreatedAt { get; set; }
-    public string? CreatedBy { get; set; }
-    public DateTimeOffset? UpdatedAt { get; set; }
-    public string? UpdatedBy { get; set; }
-    public DateTimeOffset? DeletedAt { get; set; }
-    public string? DeletedBy { get; set; }
-}
