--- DevSync.PocPro/src/core/DevSync.PocPro.Domain/DevSync.PocPro.Domain.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="Tenants\Enums\" />
-  </ItemGroup>
-
-</Project>
