--- DevSync.PocPro/src/core/DevSync.PocPro.Domain/Exceptions/DomainExceptions.cs
@@ -1,3 +1,0 @@
-namespace DevSync.PocPro.Domain.Exceptions;
-
-public class DomainExceptions(string message): Exception(message);
