--- DevSync.PocPro/src/core/DevSync.PocPro.Domain/GlobalUsings.cs
@@ -1,5 +1,0 @@
-// Global using directives
-
-global using DevSync.PocPro.Domain.Abstractions;
-global using DevSync.PocPro.Domain.Exceptions;
-global using DevSync.PocPro.Domain.Tenants.ValueObjects;
