--- DevSync.PocPro/src/infrustracture/DevSync.PocPro.Persistence/DI/RegisterPersistenceDependencies.cs
@@ -1,9 +1,0 @@
-namespace DevSync.PocPro.Persistence.DI;
-
-public static class RegisterPersistenceDependencies
-{
-    public static IServiceCollection AddPersistence(this IServiceCollection services)
-    {
-        return services;
-    }
-}
