--- DevSync.PocPro/src/infrustracture/DevSync.PocPro.Persistence/Data/TenantManagement/TenantsDbContext.cs
@@ -1,42 +1,0 @@
-using DevSync.PocPro.Domain.Abstractions;
-
-namespace DevSync.PocPro.Persistence.Data.TenantManagement;
-
-public class TenantsDbContext : DbContext
-{
-    private readonly IHttpContextAccessor _httpContextAccessor;
-
-    public TenantsDbContext(DbContextOptions<TenantsDbContext> options, IHttpContextAccessor httpContextAccessor) : base(options)
-    {
-        _httpContextAccessor = httpContextAccessor;
-    }
-
-    public override Task<int> SaveChangesAsync(CancellationToken cancellationToken = new CancellationToken())
-    {
-        foreach (var entity in ChangeTracker.Entries<IEntity>())
-        {
-            switch (entity.State)
-            {
-                case EntityState.Added:
-                    entity.Entity.CreatedBy = _httpContextAccessor.HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-                    entity.Entity.CreatedAt = DateTime.UtcNow;
-                    break;
-                case EntityState.Modified:
-                    entity.Entity.UpdatedAt = DateTime.UtcNow;
-                    entity.Entity.UpdatedBy = _httpContextAccessor.HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-                    break;
-            }
-        }
-        
-        return base.SaveChangesAsync(cancellationToken);
-    }
-    
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        modelBuilder.ApplyConfigurationsFromAssembly(typeof(TenantsDbContext).Assembly);
-        base.OnModelCreating(modelBuilder);
-    }
-
-    public DbSet<Tenant> Tenants => Set<Tenant>();
-    public DbSet<ApplicationUser> ApplicationUsers => Set<ApplicationUser>();
-}
