--- DevSync.PocPro/src/infrustracture/DevSync.PocPro.Persistence/GlobalUsings.cs
@@ -1,9 +1,0 @@
-// Global using directives
-
-global using System.Security.Claims;
-global using DevSync.PocPro.Domain.Tenants;
-global using DevSync.PocPro.Domain.Tenants.ValueObjects;
-global using Microsoft.AspNetCore.Http;
-global using Microsoft.EntityFrameworkCore;
-global using Microsoft.EntityFrameworkCore.Metadata.Builders;
-global using Microsoft.Extensions.DependencyInjection;
