--- DevSync.PocPro/src/presentation/DevSync.PocPro.Api/DevSync.PocPro.Api.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Aspire.Keycloak.Authentication" Version="9.0.0-preview.5.24551.3" />
-    <PackageReference Include="Aspire.Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.0" />
-    <PackageReference Include="Aspire.RabbitMQ.Client" Version="9.0.0" />
-    <PackageReference Include="Aspire.Seq" Version="9.0.0" />
-    <PackageReference Include="FastEndpoints" Version="6.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.4" />
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.3" />
-    <PackageReference Include="Scalar.AspNetCore" Version="2.2.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\infrustracture\DevSync.PocPro.Persistence\DevSync.PocPro.Persistence.csproj" />
-    <ProjectReference Include="..\DevSync.PocPro.Aspire\DevSync.PocPro.Aspire.ServiceDefaults\DevSync.PocPro.Aspire.ServiceDefaults.csproj" />
-  </ItemGroup>
-
-</Project>
