--- DevSync.PocPro/src/presentation/DevSync.PocPro.Api/Program.cs
@@ -1,7 +1,0 @@
-using DevSync.PocPro.Api.DI;
-
-var builder = WebApplication.CreateBuilder(args);
-
-var app = builder.AddServices().AddPipeline();
-
-app.Run();
