--- DevSync.PocPro/src/presentation/DevSync.PocPro.Aspire/DevSync.PocPro.Aspire.AppHost/Properties/launchSettings.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/launchsettings.json",
-  "profiles": {
-    "https": {
-      "commandName": "Project",
-      "dotnetRunMessages": true,
-      "launchBrowser": true,
-      "applicationUrl": "https://localhost:17131;http://localhost:15204",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DOTNET_ENVIRONMENT": "Development",
-        "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "https://localhost:21134",
-        "DOTNET_RESOURCE_SERVICE_ENDPOINT_URL": "https://localhost:22103"
-      }
-    },
-    "http": {
-      "commandName": "Project",
-      "dotnetRunMessages": true,
-      "launchBrowser": true,
-      "applicationUrl": "http://localhost:15204",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DOTNET_ENVIRONMENT": "Development",
-        "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:19171",
-        "DOTNET_RESOURCE_SERVICE_ENDPOINT_URL": "http://localhost:20234"
-      }
-    }
-  }
-}
